Move beforeAll code to JqwikMicronautExtension
